pyproject.toml
[tool.poetry]
name = "team-manager"
version = "0.8.0"
description = "Сервис автоматического управления code-review"
authors = ["Dmitriy Reznichenko <drezn@a-fin.tech>"]
readme = "README.md"
packages = [{ include = "team_manager" }]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"
pydantic = {extras = ["email"], version = "^1.10.4"}
python-gitlab = "^3.12.0"
loguru = "^0.6.0"
mattermostdriver = "^7.3.2"
fastapi = "^0.89.1"
uvicorn = { extras = ["standard"], version = "^0.20.0" }
fastapi-utils = "^0.2.1"
jinja2 = "^3.1.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
